+++ test_runner/src/main/kotlin/ftl/args/yml/MissingParameterErrorConverter.kt
@@ -2,13 +2,13 @@ package ftl.args.yml
 
 import com.fasterxml.jackson.databind.JsonNode
 import ftl.args.yml.errors.ConfigurationErrorMessageBuilder
-import ftl.util.FlankConfigurationException
+import ftl.util.FlankFatalError
 
 fun convertConfigurationErrorExceptions(missingParameterError: Exception, yaml: JsonNode): Throwable {
     val errorMessageBuilder = ConfigurationErrorMessageBuilder
     val errorMessage = missingParameterError.message
     return if (errorMessage != null) {
-        FlankConfigurationException(errorMessageBuilder(errorMessage, yaml))
+        FlankFatalError(errorMessageBuilder(errorMessage, yaml))
     } else {
         missingParameterError
     }